requirements.txt
google-api-python-client==2.52.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
